+++ services/projects/index.js
@@ -136,6 +136,10 @@ async function query (attrs, opts = {}) {
     }
   }
 
+  if (attrs.organization_id) {
+    where.organization_id = attrs.organization_id
+  }
+
   const method = (!!attrs.limit || !!attrs.offset) ? 'findAndCountAll' : 'findAll' // don't use findAndCountAll if we don't need to limit and offset
   return models.Project[method]({
     where,