+++ services/streams/index.js
@@ -144,6 +144,10 @@ async function query (attrs, opts = {}) {
     }
   }
 
+  if (attrs.projects) {
+    where.project_id = attrs.projects
+  }
+
   const method = (!!attrs.limit || !!attrs.offset) ? 'findAndCountAll' : 'findAll' // don't use findAndCountAll if we don't need to limit and offset
   return models.Stream[method]({
     where,